refactor:#70 아바타스텍 유저정보 클라이언트에서 호출하는거 서버에서 패치하고 프랍스로 전달해주는 로직으로 리팩터링
